requirements.txt
Django==2.0.4
Pillow==5.1.0
astroid==1.6.3
autopep8==1.3.5
coverage==4.5.1
django-abstract-related-model==0.1.1
django_extensions==1.8.1
django_singleton_model==0.2.3
djangorestframework-jwt==1.11.0
djangorestframework==3.8.1
coreapi==2.3.3
factory-boy==2.9.2
flake8==3.5.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.6.0
pylint-django==0.11.1
pylint-plugin-utils==0.2.6
pylint==1.8.4
pytz==2018.3
six==1.11.0
wrapt==1.10.11
pyaml==17.12.1
